requirements.txt
python-dotenv
discord.py
Pillow
python-dateutil
requests
pyquery
pytz